Register Contoller and  Postman bug fixes

+++ Backend Project/src/utils/cloudinary.js
@@ -1,15 +1,17 @@
 import { v2 as cloudinary } from "cloudinary";
 import fs from "fs";
 
+
+
 //configuration code to connect project with cloudINARY SERVICES
 cloudinary.config({
-  cloud_name: "process.env.CLOUDINARY_CLOUD_NAME",
-  api_key: "process.env.CLOUDINARY_API_KEY",
-  api_secret: "process.env.CLOUDINARY_API_SECRET",
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+  api_key: process.env.CLOUDINARY_API_KEY,
+  api_secret: process.env.CLOUDINARY_API_SECRET,
 });
 
 //Method to upload file in cloudinary
-const uoloadCloudinary = async (localfilePath) => {
+const uploadCloudinary = async (localfilePath) => {
   
   try {
     //file uploading
@@ -18,16 +20,20 @@ const uoloadCloudinary = async (localfilePath) => {
       resource_type: "auto",
     });
     // file uploaded successfully
-    console.log("file is successfully uploaded to cloudinary", response.url);
+    //console.log("file is successfully uploaded to cloudinary", response.url);
+    fs.unlinkSync(localfilePath);
     return response;
   } 
   
   catch (error) {
-    //fs is filesystem in node js that is used to remove temporarily stored file on server because file 
-    //uploading operatiom got failed
-    fs.unlinkSync(localfilePath)
-
+    // THIS LINE IS KEY: It will print the real reason in your terminal
+    console.error("CLOUDINARY UPLOAD ERROR:", error.message); 
+    
+    if (fs.existsSync(localfilePath)) {
+        fs.unlinkSync(localfilePath);
+    }
+    return null;
   }
 };
 
-export {uoloadCloudinary}
+export {uploadCloudinary}
